frontend backend connection done

+++ controllers/authenticate.js
@@ -1,17 +1,18 @@
 const User = require('../models/User');
 const checkCredentials = require('../scraper/login');
-const {UnauthenticatedError} = require('../errors');
+const {UnauthenticatedError, BadRequestError} = require('../errors');
 const { StatusCodes } = require('http-status-codes');
 
 const register = async(req, res) => {
-    const userPassword = req.body.password;
+    const {name, contactNumber, email, password:userPassword} = req.body;
+    if (!name || !contactNumber || !email || !userPassword) {
+        throw new BadRequestError('Please provide necessary details!')
+    }
     const user = await User.create({...req.body})
-    console.log(user)
     const isValid = await checkCredentials(user.email, userPassword).then(response => response)
-    console.log(isValid)
     if (!isValid) {
         await User.findOneAndDelete({email : user.email})
-        throw new UnauthenticatedError('Invalid Credentials!')
+        throw new UnauthenticatedError('The email or password provided is incorrect')
     }
     res.status(StatusCodes.CREATED).json({msg: `You have been subscribed to the attendance chatbot and will be receiving updates on ${user.contactNumber}`})
 }
